Separate RGBGFX warnings/errors from main options
